[docker] create light faster server client
