+++ docker/Dockerfile.server
@@ -1,3 +1,11 @@
-FROM waika_project_image
+ARG NODE_VERSION=20.12.2
+ARG SERVER_PORT=3001
 
-CMD ["npm", "run", "dev:server"]
+FROM node:$NODE_VERSION-alpine
+
+WORKDIR /app/server
+COPY packages/server/dist/ ./
+
+RUN npm install -g nodemon
+
+CMD ["nodemon", "server.js"]
