--- packages/server/nodemon.json
@@ -1,6 +1,0 @@
-{
-  "verbose": false,
-  "ignore": ["*.test.ts", "node_modules"],
-  "watch": ["index.ts", "src/**"],
-  "ext": "ts,json"
-}
