+++ packages/server/scripts/esbuild.prod.ts
@@ -1,0 +1,11 @@
+import { build } from 'esbuild';
+import { esBuildConfig } from './esbuild.config';
+
+async function buildProd() {
+  await build(esBuildConfig);
+}
+
+buildProd().catch((error) => {
+  console.error(error);
+  process.exit(1);
+});
